Initial commit of VoteChange
- Create VoteChange.java, the plugin class which initializes commands and registers enabling and disabling of the function.

- Create CallVote.java, a command which allows users to call a vote requesting to execute a command. If a majority of 2/3 or greater of voting users vote yes, then the vote is approved and the user is granted momentary server operator status to execute the command.

- Create Vote.java, a command which allows users to vote either "yes" or "no" only when a vote is created in CallVote.

This still has the flaw of making users server operator, although only for a moment, while they execute the command which gained a majority of "yes" votes. The intention is to change this to grant them permission only for the command they received access to.

Another flaw exists in CallVote.java, where only one vote can be called at a time. The intention is to implement a queue of CallVote objects if there is already a vote in progress, and dequeueing and executing the top CallVote command when the prior vote is completed.

Another potential flaw is the notion that 2/3 of voting users must vote yes. If 2/3 of online users are required to vote yes, it is possible that users who always abstain are effectively "no" votes. On the other hand, if people consistently neglect voting then a small group of active voters can execute commands with leniency. For the time being, voting users will suffice.

A final flaw in development is resolving the issue that users may be granted too much leniency in which commands are allowed to be executed. A config.yml file will be created, granting server owners the ability to blacklist commands from voting.